Move type classes to a separate package
