--- src/main/kotlin/sigma/Type.kt
@@ -1,5 +1,0 @@
-package sigma
-
-interface Type
-
-object IntType : Type
