+++ src/main/kotlin/sigma/expressions/TypeReference.kt
@@ -1,8 +1,8 @@
 package sigma.expressions
 
-import sigma.values.Symbol
-import sigma.Type
 import sigma.TypeContext
+import sigma.types.Type
+import sigma.values.Symbol
 
 data class TypeReference(
     val referee: Symbol,
