+++ src/main/kotlin/sigma/types/FunctionType.kt
@@ -1,0 +1,6 @@
+package sigma.types
+
+data class FunctionType(
+    val argumentType: Type,
+    val imageType: Type,
+) : Type
