+++ src/main/kotlin/sigma/types/Type.kt
@@ -1,0 +1,5 @@
+package sigma.types
+
+sealed interface Type
+
+object IntType : Type
